refactor: resolve the linter problems
