+++ delivery/websocket/configs/config.go
@@ -28,28 +28,30 @@ type Config struct {
 	limitation *Limitation
 }
 
-// GetURL safely retrieves the URL
+// GetURL safely retrieves the URL.
 func (c *Config) GetURL() *url.URL {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
+
 	return c.url
 }
 
-// SetURL safely sets the URL
+// SetURL safely sets the URL.
 func (c *Config) SetURL(u *url.URL) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	c.url = u
 }
 
-// GetLimitation safely retrieves the Limitation
+// GetLimitation safely retrieves the Limitation.
 func (c *Config) GetLimitation() *Limitation {
 	c.mu.RLock()
 	defer c.mu.RUnlock()
+
 	return c.limitation
 }
 
-// SetLimitation safely sets the Limitation
+// SetLimitation safely sets the Limitation.
 func (c *Config) SetLimitation(l *Limitation) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
