+++ delivery/websocket/event_handler.go
@@ -2,18 +2,19 @@ package websocket
 
 import (
 	"fmt"
-	"github.com/dezh-tech/immortal/delivery/websocket/configs"
 	"slices"
 	"strconv"
 	"strings"
 	"time"
 
+	"github.com/dezh-tech/immortal/delivery/websocket/configs"
 	"github.com/dezh-tech/immortal/infrastructure/redis"
 	"github.com/dezh-tech/immortal/pkg/logger"
 	"github.com/dezh-tech/immortal/pkg/utils"
 	"github.com/dezh-tech/immortal/types"
 	"github.com/dezh-tech/immortal/types/filter"
 	"github.com/dezh-tech/immortal/types/message"
+
 	"github.com/gorilla/websocket"
 )
