+++ infrastructure/grpc_client/proto/config.proto
@@ -10,17 +10,17 @@ service Parameters {
 }
 
 message Limitations {
-    int32 max_message_length = 1;        
-    int32 max_subscriptions = 2;                    
-    int32 max_subid_length = 4;        
-    int32 min_pow_difficulty = 5;          
-    bool auth_required = 6;              
-    bool payment_required = 7;             
-    bool restricted_writes = 8;         
-    int32 max_event_tags = 9;            
-    int32 max_content_length = 10;    
+    int32 max_message_length = 1;
+    int32 max_subscriptions = 2;
+    int32 max_subid_length = 4;
+    int32 min_pow_difficulty = 5;
+    bool auth_required = 6;
+    bool payment_required = 7;
+    bool restricted_writes = 8;
+    int32 max_event_tags = 9;
+    int32 max_content_length = 10;
     int64 created_at_lower_limit = 11;
-    int64 created_at_upper_limit = 12;    
+    int64 created_at_upper_limit = 12;
     uint32 default_query_limit = 13;    
     uint32 max_query_limit = 14;           
 }
@@ -28,8 +28,8 @@ message Limitations {
 message GetParametersRequest {}
 
 message GetParametersResponse {
-    Limitations limitations = 1;          
-    string url = 2;                       
+    Limitations limitations = 1;
+    string url = 2;
 }
 
 message UpdateParametersResponse {
